server/requirements.txt
gevent==21.1.2
greenlet==1.0.0
tornado==6.1
websocket==0.2.1
zope.event==4.5.0
zope.interface==5.2.0
